+++ src/main/java/com/jettech/service/Impl/QualitySuiteServiceImpl.java
@@ -5,11 +5,18 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import javax.transaction.Transactional;
+
 import com.jettech.entity.QualityRule;
+import com.jettech.entity.QualityTestQuery;
 import com.jettech.repostory.QualityRuleRepository;
+import com.jettech.repostory.QualityTestQueryRepository;
 import com.jettech.vo.QualitySuiteVO;
 import com.jettech.vo.ResultVO;
 import com.jettech.vo.StatusCode;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -23,14 +30,14 @@ import com.jettech.service.IQualitySuiteService;
 
 @Service
 public class QualitySuiteServiceImpl implements IQualitySuiteService {
+	private static Logger log = LoggerFactory.getLogger(QualitySuiteServiceImpl.class);
 
-	//@Autowired
-	//QualitySuiteRepository qualitySuiteRepository;
 	@Autowired
 	QualitySuiteRepository qualitySuiteRepository;
 	@Autowired
 	QualityRuleRepository qualityRuleRepository;
-
+	@Autowired
+    QualityTestQueryRepository qualityTestQueryRepository;
 	@Override
 	public List<QualitySuite> findAll() {
 		return qualitySuiteRepository.findAll();
@@ -92,13 +99,26 @@ public class QualitySuiteServiceImpl implements IQualitySuiteService {
 	}
 
 	@Override
+	@Transactional
 	public ResultVO delete(int id) {
-		QualitySuite qualitySuite = qualitySuiteRepository.findById(id).get();;
-		qualitySuite.setQualityRules(null);
-		qualitySuiteRepository.save(qualitySuite);
-		qualitySuiteRepository.deleteById(id);
 		Map<String, Object> map = new HashMap<String, Object>();
-		return new ResultVO(true, StatusCode.OK, "删除成功", map);
+		try {
+			QualitySuite qualitySuite = qualitySuiteRepository.findById(id).get();;
+			List<QualityTestQuery> qualityTestQuerys=qualityTestQueryRepository.findByQualitySuiteId(qualitySuite.getId());
+			if(qualityTestQuerys.isEmpty()){
+				qualitySuite.setQualityRules(null);
+				qualitySuiteRepository.save(qualitySuite);
+				qualitySuiteRepository.deleteById(id);
+				return new ResultVO(true, StatusCode.OK, "删除成功", map);
+			}else{
+				return new ResultVO(false, StatusCode.ERROR, "此集合关联案例禁止删除");
+			}
+		} catch (Exception e) {
+			log.info("异常为"+e);
+			return new ResultVO(false, StatusCode.ERROR, "删除失败");
+
+		}
+		
 	}
 
 	@Override
@@ -127,21 +147,7 @@ public class QualitySuiteServiceImpl implements IQualitySuiteService {
 		Page<QualitySuite> pageQualitySuite = qualitySuiteRepository.findAll(pageRequest);
 		List<QualitySuiteVO> qualitySuiteVos=new ArrayList<QualitySuiteVO>();
 		for(QualitySuite qualitySuite : pageQualitySuite){
-			QualitySuiteVO qualitySuiteVo=new QualitySuiteVO();
-			BeanUtils.copyProperties(qualitySuite, qualitySuiteVo);
-			List<QualityRule> qualityRules=new ArrayList<QualityRule>();
-			StringBuilder qualityRuleIds=new StringBuilder();
-			qualityRules=qualitySuite.getQualityRules();
-			for (int i = 0; i < qualityRules.size(); i++) {
-				// 当循环到最后一个的时候 就不添加逗号,
-				if (i == qualityRules.size() - 1) {
-					qualityRuleIds.append(qualityRules.get(i).getId());
-				} else {
-					qualityRuleIds.append(qualityRules.get(i).getId());
-					qualityRuleIds.append(",");
-				}
-			}
-			qualitySuiteVo.setQualityRuleIds(qualityRuleIds.toString());
+			QualitySuiteVO qualitySuiteVo=new QualitySuiteVO(qualitySuite);
 			qualitySuiteVos.add(qualitySuiteVo);
 		}
